requirements.txt
streamlit>=1.30.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
anthropic>=0.69.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-anthropic>=1.0.0
langchain-community>=0.3.0
sqlalchemy>=2.0.0